# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for a private upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
components:
...
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
...